requirements.txt
facenet-pytorch==2.6.0
fastapi==0.116.1
mediapipe==0.10.21
mysql-connector-python==9.4.0
numpy==1.26.4
opencv-python==4.10.0.82
python-dotenv==1.1.1
python-multipart==0.0.20
scikit-image==0.25.2
scikit-learn==1.7.1
SQLAlchemy==2.0.43
torch==2.2.2
uvicorn==0.35.0